feat: implement CardViewer component for TabiPayCardPage with download and share functionality
